requirements.txt
requests
PyYAML
Jinja2
PyGithub
